Refactor Caesar Cipher functionality into one class
Consolidate encryption and decryption methods into a single `CaesarCipher` class. Remove `CaesarCipherEncode` and `CaesarCipherDecode` classes. Update project references and method calls in `Program.cs` to use the new structure.
