+++ EncryptionTest.csproj
@@ -43,8 +43,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CaesarCipherDecode.cs" />
-    <Compile Include="CaesarCipherEncode.cs" />
+    <Compile Include="CaesarCipher.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
